timeToHide settings/options property updated
Added the possibility of putting the 'false' value to options.timeToHide. That resulted in having the possibility for a user to not specify the exact milliseconds value to timeToHide and therefore getting the infinite spinner. Useful in situations where the user needs to wait longer time for some content to be ready in the background before fading out the spinner.